emon {network}	R Documentation

Interorganizational Search and Rescue Networks (Drabek et al.)

Description

Drabek et al. (1981) provide seven case studies of emergent multi-organizational networks (EMONs) in the context of search and rescue (SAR) activities. Networks of interaction frequency are reported, along with several organizational attributes.

Usage

data(emon)

Format

A list of 7 network objects:

[[1]] 	 Cheyenne 	 network 	 Cheyenne SAR EMON

[[2]] 	 HurrFrederic 	 network 	 Hurricane Frederic SAR EMON

[[3]] 	 LakePomona 	 network 	 Lake Pomona SAR EMON

[[4]] 	 MtSi 	 network 	 Mt. Si SAR EMON

[[5]] 	 MtStHelens 	 network 	 Mt. St. Helens SAR EMON

[[6]] 	 Texas 	 network 	 Texas Hill Country SAR EMON

[[7]] 	 Wichita 	 network 	 Wichita Falls SAR EMON

Each network has one edge attribute:

Frequency 	 numeric 	 Interaction frequency (1-4; 1=most frequent)

Each network also has 8 vertex attributes:

Command.Rank.Score 	 numeric 	 Mean rank in the command structure

Decision.Rank.Score 	 numeric 	 Mean rank in the decision process

Formalization 	 numeric 	 Degree of formalization

Location 	 character 	 Location code

Paid.Staff 	 numeric 	 Number of paid staff

Sponsorship 	 character 	 Sponsorship type

vertex.names 	 character 	 Organization name

Volunteer.Staff 	 numeric 	 Number of volunteer staff

Details

All networks collected by Drabek et al. reflect reported frequency of organizational interaction during the search and rescue effort; the (i,j) edge constitutes i's report regarding interaction with j, with non-adjacent vertices reporting no contact. Frequency is rated on a four-point scale, with 1 indicating the highest frequency of interaction. (Response options: 1=“continuously”, 2=“about once an hour”, 3=“every few hours”, 4=“about once a day or less”) This is stored within the "Frequency" edge attribute.

For each network, several covariates are recorded as vertex attributes:

Command.Rank.Score

Mean (reversed) rank for the prominence of each organization in the command structure of the response, as judged by organizational informants.

Decision.Rank.Score

Mean (reversed) rank for the prominence of each organization in decision making processes during the response, as judged by organizational informants.

Formalization

An index of organizational formalization, ranging from 0 (least formalized) to 4 (most formalized).

Localization

For each organization, "L" if the organization was sited locally to the impact area, "NL" if the organization was not sited near the impact area, "B" if the organization was sited at both local and non-local locations.

Paid.Staff

Number of paid staff employed by each organization at the time of the response.

Sponsorship

The level at which each organization was sponsored (e.g., "City", "County", "State", "Federal", and "Private").

vertex.names

The identity of each organization.

Volunteer.Staff

Number of volunteer staff employed by each organization at the time of the response.

Note that where intervals were given by the original source, midpoints have been substituted. For detailed information regarding data coding and procedures, see Drabek et al. (1981).

Source

Drabek, T.E.; Tamminga, H.L.; Kilijanek, T.S.; and Adams, C.R. (1981). Data from Managing Multiorganizational Emergency Responses: Emergent Search and Rescue Networks in Natural Disaster and Remote Area Settings. Program on Technology, Environment, and Man Monograph 33. Institute for Behavioral Science, University of Colorado.

See Also

network

Examples

data(emon)   #Load the emon data set

#Plot the EMONs
par(mfrow=c(3,3))
for(i in 1:length(emon))
  plot(emon[[i]],main=names(emon)[i],edge.lwd="Frequency")

[Package network version 1.13.0.1 Index]