has.edges {network}	R Documentation

Determine if specified vertices of a network have any edges (are not isolates)

Description

Returns a logical value for each specified vertex, indicating if it has any incident (in or out) edges. Checks all vertices by default

Usage

has.edges(net, v = seq_len(network.size(net)))

Arguments

net
a network object to be queried

v
integer vector of vertex ids to check

Value

returns a logical vector with the same length as v, with TRUE if the vertex is involved in any edges, FALSE if it is an isolate.

Author(s)

skyebend

Examples

test<-network.initialize(5)
test[1,2]<-1
has.edges(test)
has.edges(test,v=5)

[Package network version 1.14-377 Index]