network-internal {network}	R Documentation

Internal Network Package Functions

Description

Internal network functions.

Usage

is.discrete(x)
is.discrete.character(x)
is.discrete.numeric(x)
mixingmatrix(object, ...)
## S3 method for class 'network'
mixingmatrix(object, attrname, ...)
networkOperatorSetup(x,y=NULL)

Arguments

object
a network or some other data structure for which a mixing matrix is meaningful.

x
an object to be designated either discrete or continuous, or a network.

attrname
a vertex attribute name.

y
a network or something coercible to one.

...
further arguments passed to or used by methods.

Details

Most of these are not to be called by the user.

See Also

network

[Package network version 1.14-377 Index]