network.initialize {network}	R Documentation

Initialize a Network Class Object

Description

Create and initialize a network object with n vertices.

Usage

network.initialize(n, directed = TRUE, hyper = FALSE, loops = FALSE, 
    multiple = FALSE, bipartite = FALSE)

Arguments

n
the number of vertices to initialize

directed
logical; should edges be interpreted as directed?

hyper
logical; are hyperedges allowed?

loops
logical; should loops be allowed?

multiple
logical; are multiplex edges allowed?

bipartite
count; should the network be interpreted as bipartite? If present (i.e., non-NULL) it is the count of the number of actors in the first mode of the bipartite network. In this case, the overall number of vertices is equal to the number of 'actors' (first mode) plus the number of ‘events’ (second mode), with the vertex.ids of all actors preceeding all events. The edges are then interpreted as nondirected.

Details

Generally, network.initialize is called by other constructor functions as part of the process of creating a network.

Value

An object of class network

Author(s)

Carter T. Butts buttsc@uci.edu

References

Butts, C. T. (2008). “network: a Package for Managing Relational Data in R.” Journal of Statistical Software, 24(2). http://www.jstatsoft.org/v24/i02/

See Also

network, as.network.matrix

Examples

g<-network.initialize(5)  #Create an empty graph on 5 vertices

[Package network version 1.13.0.1 Index]